requirements.txt
# External packages
cryptography
asn1crypto
pytz
pyyaml
appdirs
pyserial
python-jose[cryptography]

# Microchip packages
pykitcommander>=2.3.2*
pyedbglib>=2.17
